add bing search script
We use bing-search.py to search Bing for TSE IDs and get polling data
from news reports. This is necessary for pollsters that do not share
their data on their websites, or that do not answer our requests for
polls.

In the future, I plan to integrate bing-search.py better with
bing_utils.R, to avoid manual data handling between the two. Meanwhile,
they are separate.