code: create decent generator for 2020 polls
It's October, and we finally have the final candidate list for the 2020
elections. This meant that we had to start including estimulada polls in
our outputs.

This commit introduces parse_2020_2.R, which generates 2020 polls
roughly in the same format that we generated for the pre-2020 years. It
does most of the sanity checks, the inappropriate scenario removals,
etc.