data: add 2020 poll scaffolding, and new polls
Before October, we used an HTML scraper to crawl new polls from the TSE
registry. Now, the actual pesqEle files are generated by the court every
day, and they provide a more reliable registry database. We add it to
the repository in this commit.

We also add a number of 2020 polls -- some that we missed prior to
Octobre; and some new ones from the last few days.