code: allow importation of HTML Registry data
Not too unexpectedly, the TSE decided to stop supplying poll registry
data for a few states in the past few days. We can't go without having
our polls, so we resorted to a backup of a sort.

We'll use the pesqEle HTML parser to extract polls for the
states/periods that we want, and export them into special files in
data/tse/*. Here, we change the code to translate these "special" files
into the TSE CSV format. That way, we seamlessly integrate that into
our existsing logic for generating polls.